Fix cross-platform build warnings and MSVC compatibility
Make the C/CUDA library and tests build cleanly on Windows (MSVC),
macOS, and Linux without changing kernel math.

- CMakeLists: gate GCC-only flags behind NOT MSVC, make OpenMP and
  CUDA flags platform-aware, detect newer MSVC and pass
  -allow-unsupported-compiler, add --expt-relaxed-constexpr and
  -Wno-deprecated-gpu-targets, co-locate the DLL with test exes on
  Windows, drop the hardcoded --std=c++11 from test targets.
- rnnt_helper.h: use std::numeric_limits<T>::infinity() to avoid the
  float-overflow warning, type the threshold constant as float.
- cpu_rnnt.h: use ProbT(1) instead of a double literal.
- tests: add f suffixes to float literals, cast size() to int, use
  0.0f accumulate identities to silence MSVC C4244/C4267/C4305.

+++ tests/test_cpu.cpp
@@ -15,15 +15,15 @@ bool small_test() {
     const int T = 2;
     const int U = 3;
 
-    std::vector<float> acts = {0.1, 0.6, 0.1, 0.1, 0.1, 0.1, 
-                                0.1, 0.6, 0.1, 0.1, 0.1, 0.1, 
-                                0.2, 0.8, 0.1, 0.1, 0.6, 0.1, 
-                                0.1, 0.1, 0.1, 0.1, 0.2, 0.1, 
-                                0.1, 0.7, 0.1, 0.2, 0.1, 0.1};
+    std::vector<float> acts = {0.1f, 0.6f, 0.1f, 0.1f, 0.1f, 0.1f,
+                                0.1f, 0.6f, 0.1f, 0.1f, 0.1f, 0.1f,
+                                0.2f, 0.8f, 0.1f, 0.1f, 0.6f, 0.1f,
+                                0.1f, 0.1f, 0.1f, 0.1f, 0.2f, 0.1f,
+                                0.1f, 0.7f, 0.1f, 0.2f, 0.1f, 0.1f};
     std::vector<float> log_probs(acts.size());
     softmax(acts.data(), alphabet_size, B * T * U, log_probs.data(), true);
 
-    float expected_score = 4.495666;
+    float expected_score = 4.495666f;
 
     std::vector<int> labels = {1, 2};
     std::vector<int> label_lengths = {2};
@@ -55,14 +55,14 @@ bool small_test() {
                                     label_lengths.data(),
                                     lengths.data(),
                                     alphabet_size,
-                                    lengths.size(),
+                                    (int)lengths.size(),
                                     &score,
                                     rnnt_cpu_workspace,
                                     options),
                    "Error: compute_rnnt_loss in small_test");
 
     free(rnnt_cpu_workspace);
-    const float eps = 1e-4;
+    const float eps = 1e-4f;
 
     const float lb = expected_score - eps;
     const float ub = expected_score + eps;
@@ -76,33 +76,33 @@ bool options_test() {
     const int L = 3;
     const int minibatch = 2;
 
-    std::vector<float> acts = {0.065357, 0.787530, 0.081592, 0.529716, 0.750675, 0.754135, 
-                                0.609764, 0.868140, 0.622532, 0.668522, 0.858039, 0.164539, 
-                                0.989780, 0.944298, 0.603168, 0.946783, 0.666203, 0.286882, 
-                                0.094184, 0.366674, 0.736168, 0.166680, 0.714154, 0.399400, 
-                                0.535982, 0.291821, 0.612642, 0.324241, 0.800764, 0.524106, 
-                                0.779195, 0.183314, 0.113745, 0.240222, 0.339470, 0.134160, 
-                                0.505562, 0.051597, 0.640290, 0.430733, 0.829473, 0.177467, 
-                                0.320700, 0.042883, 0.302803, 0.675178, 0.569537, 0.558474, 
-                                0.083132, 0.060165, 0.107958, 0.748615, 0.943918, 0.486356, 
-                                0.418199, 0.652408, 0.024243, 0.134582, 0.366342, 0.295830, 
-                                0.923670, 0.689929, 0.741898, 0.250005, 0.603430, 0.987289, 
-                                0.592606, 0.884672, 0.543450, 0.660770, 0.377128, 0.358021};
+    std::vector<float> acts = {0.065357f, 0.787530f, 0.081592f, 0.529716f, 0.750675f, 0.754135f,
+                                0.609764f, 0.868140f, 0.622532f, 0.668522f, 0.858039f, 0.164539f,
+                                0.989780f, 0.944298f, 0.603168f, 0.946783f, 0.666203f, 0.286882f,
+                                0.094184f, 0.366674f, 0.736168f, 0.166680f, 0.714154f, 0.399400f,
+                                0.535982f, 0.291821f, 0.612642f, 0.324241f, 0.800764f, 0.524106f,
+                                0.779195f, 0.183314f, 0.113745f, 0.240222f, 0.339470f, 0.134160f,
+                                0.505562f, 0.051597f, 0.640290f, 0.430733f, 0.829473f, 0.177467f,
+                                0.320700f, 0.042883f, 0.302803f, 0.675178f, 0.569537f, 0.558474f,
+                                0.083132f, 0.060165f, 0.107958f, 0.748615f, 0.943918f, 0.486356f,
+                                0.418199f, 0.652408f, 0.024243f, 0.134582f, 0.366342f, 0.295830f,
+                                0.923670f, 0.689929f, 0.741898f, 0.250005f, 0.603430f, 0.987289f,
+                                0.592606f, 0.884672f, 0.543450f, 0.660770f, 0.377128f, 0.358021f};
     std::vector<float> log_probs(acts.size());
     softmax(acts.data(), alphabet_size, minibatch * T * L, log_probs.data(), true);
 
-    std::vector<float> expected_grads = {-0.432226, -0.567774, 0        , -0.365650, 0        , -0.202123, 
-                                        -0.202123, 0        , 0        , -0.165217, -0.267010, 0        , 
-                                        -0.394365, 0        , -0.238294, -0.440418, 0        , 0        , 
-                                        -0.052130, -0.113087, 0        , -0.183138, 0        , -0.324314, 
-                                        -0.764732, 0        , 0        , 0        , -0.052130, 0        , 
-                                        0        , 0        , -0.235268, -1       , 0        , 0        , 
-                                        -0.716142, -0.283858, 0        , -0.183829, -0.100028, 0        , 
-                                        -0.100028, 0        , 0        , -0.411218, -0.304924, 0        , 
-                                        -0.329576, -0.159178, 0        , -0.259206, 0        , 0        , 
-                                        -0.116076, -0.295142, 0        , -0.286533, -0.338184, 0        , 
-                                        -0.597390, 0        , 0        , 0        , -0.116076, 0        , 
-                                        0        , -0.402610, 0        , -1       , 0        , 0        };
+    std::vector<float> expected_grads = {-0.432226f, -0.567774f, 0.0f, -0.365650f, 0.0f, -0.202123f,
+                                        -0.202123f, 0.0f, 0.0f, -0.165217f, -0.267010f, 0.0f,
+                                        -0.394365f, 0.0f, -0.238294f, -0.440418f, 0.0f, 0.0f,
+                                        -0.052130f, -0.113087f, 0.0f, -0.183138f, 0.0f, -0.324314f,
+                                        -0.764732f, 0.0f, 0.0f, 0.0f, -0.052130f, 0.0f,
+                                        0.0f, 0.0f, -0.235268f, -1.0f, 0.0f, 0.0f,
+                                        -0.716142f, -0.283858f, 0.0f, -0.183829f, -0.100028f, 0.0f,
+                                        -0.100028f, 0.0f, 0.0f, -0.411218f, -0.304924f, 0.0f,
+                                        -0.329576f, -0.159178f, 0.0f, -0.259206f, 0.0f, 0.0f,
+                                        -0.116076f, -0.295142f, 0.0f, -0.286533f, -0.338184f, 0.0f,
+                                        -0.597390f, 0.0f, 0.0f, 0.0f, -0.116076f, 0.0f,
+                                        0.0f, -0.402610f, 0.0f, -1.0f, 0.0f, 0.0f};
     // Calculate the expected scores analytically
     std::vector<double> expected_scores(2);
     expected_scores[0] = 4.2806528590890736;
@@ -138,7 +138,7 @@ bool options_test() {
                                     label_lengths.data(),
                                     lengths.data(),
                                     alphabet_size,
-                                    lengths.size(),
+                                    (int)lengths.size(),
                                     scores.data(),
                                     rnnt_cpu_workspace,
                                     options),
@@ -222,7 +222,7 @@ bool inf_test() {
                                     label_lengths.data(),
                                     sizes.data(),
                                     alphabet_size,
-                                    sizes.size(),
+                                    (int)sizes.size(),
                                     &cost,
                                     rnnt_cpu_workspace,
                                     options),
@@ -243,7 +243,7 @@ void numeric_grad(std::vector<float>& acts, std::vector<int>& flat_labels, std::
                 std::vector<int> sizes, int alphabet_size, int minibatch, 
                 void* rnnt_cpu_workspace, rnntOptions& options, std::vector<float>& num_grad) {
 
-    float epsilon = 1e-2;
+    float epsilon = 1e-2f;
 
     for (int i = 0; i < num_grad.size(); ++i) {
 
@@ -276,8 +276,8 @@ void numeric_grad(std::vector<float>& acts, std::vector<int>& flat_labels, std::
                                         options),
                        "Error: compute_rnnt_loss (2) in grad_check");
 
-        float costP1 = std::accumulate(costsP1.begin(), costsP1.end(), 0.);
-        float costP2 = std::accumulate(costsP2.begin(), costsP2.end(), 0.);
+        float costP1 = std::accumulate(costsP1.begin(), costsP1.end(), 0.0f);
+        float costP2 = std::accumulate(costsP2.begin(), costsP2.end(), 0.0f);
 
         acts[i] += epsilon;
         num_grad[i] = (costP1 - costP2) / (2 * epsilon);
@@ -289,13 +289,13 @@ bool grad_check(int T, int L, int alphabet_size,
                   const std::vector<std::vector<int>>& labels,
                   const std::vector<int>& sizes, float tol) {
 
-    const int minibatch = labels.size();
+    const int minibatch = (int)labels.size();
 
     std::vector<int> flat_labels;
     std::vector<int> label_lengths;
     for (const auto& l : labels) {
         flat_labels.insert(flat_labels.end(), l.begin(), l.end());
-        label_lengths.push_back(l.size());
+        label_lengths.push_back((int)l.size());
     }
 
     std::vector<float> costs(minibatch);
@@ -310,7 +310,7 @@ bool grad_check(int T, int L, int alphabet_size,
     options.batch_first = true;
 
     size_t cpu_alloc_bytes;
-    throw_on_error(get_workspace_size(T, L, sizes.size(),
+    throw_on_error(get_workspace_size(T, L, (int)sizes.size(),
                                       false,
                                       &cpu_alloc_bytes),
                    "Error: get_workspace_size in grad_check");
@@ -329,7 +329,7 @@ bool grad_check(int T, int L, int alphabet_size,
                                     options),
                    "Error: compute_rnnt_loss (0) in grad_check");
 
-    float cost = std::accumulate(costs.begin(), costs.end(), 0.);
+    float cost = std::accumulate(costs.begin(), costs.end(), 0.0f);
 
     std::vector<float> num_grad(grads.size());
 
@@ -346,8 +346,8 @@ bool grad_check(int T, int L, int alphabet_size,
 
 bool run_tests() {
     std::vector<std::tuple<int, int, int, int, float>> problem_sizes =
-       {std::make_tuple(20, 50, 15, 1, 1e-4),
-        std::make_tuple(5, 10, 5, 65, 1e-4)
+       {std::make_tuple(20, 50, 15, 1, 1e-4f),
+        std::make_tuple(5, 10, 5, 65, 1e-4f)
        };
 
     std::mt19937 gen(2);
